all: Extract authentication method to its own file
